Overcame my smooth brain
Added an initialize method to the controller that will run when the UI loads. In the initialize method I added 3 methods to have the active and complete category lists load with an all tasks button, as well as to populate the user role choice box that I forgot to add to the user creation tab, but that is there now. Also cleaned up the fxml injections to be more visibly grouped now that we are mostly done editing them